refactor: 헥사고날 아키텍처 개선 및 ID 타입 통일
- String ID에서 Long ID로 전환하여 타입 안정성 향상
- 도메인 서비스 의존성 주입 개선 (DI 패턴 적용)
- Controller try-catch 제거 (글로벌 예외 처리 활용)
- 포트와 어댑터 인터페이스 일관성 개선
- 새로운 RefreshToken 도메인 모델 추가

+++ src/main/kotlin/com/monkeys/spark/application/service/InquiryApplicationService.kt
@@ -18,7 +18,7 @@ class InquiryApplicationService(
 ) : InquiryUseCase {
     
     override fun createInquiry(command: CreateInquiryCommand): Inquiry {
-        val userId = command.userId?.let { UserId(it) }
+        val userId = command.userId?.let { UserId(it.toLong()) }
         
         val inquiry = Inquiry.create(
             userId = userId,
